+++ cards/moengage_cards_ios/lib/moengage_cards_ios.dart
@@ -40,11 +40,11 @@ class MoEngageCardsIOS extends MoEngageCardsPlatform {
   }
 
   @override
-  void setAppOpenCardsSyncListener(
+  void setSyncCompleteListener(
     CardsSyncListener cardsSyncListener,
     String appId,
   ) {
-    super.setAppOpenCardsSyncListener(cardsSyncListener, appId);
+    super.setSyncCompleteListener(cardsSyncListener, appId);
     methodChannel.invokeMethod(
       methodSetAppOpenCardsSyncListener,
       getAccountMeta(appId),