+++ cards/moengage_cards_platform_interface/lib/src/internal/callback_cache.dart
@@ -2,8 +2,9 @@ import '../../moengage_cards_platform_interface.dart';
 
 /// Callback Cache for Cards Module
 class CallbackCache {
-  /// App Open Cards Sync Listener
-  CardsSyncListener? appOpenSyncListener;
+  /// Sync Listener Cache for Card syncs performed by SDK.
+  /// Currently applicable for AppOpen Sync, Card fetch on login ..
+  CardsSyncListener? cardsSyncListener;
 
   /// Inbox Open Cards Sync Listener
   CardsSyncListener? inboxOpenSyncListener;