+++ cards/moengage_cards_platform_interface/lib/src/internal/constants.dart
@@ -93,10 +93,11 @@ const String methodNewCardsCount = 'getNewCardsCount';
 const String methodUnClickedCardsCount = 'unClickedCardsCount';
 const String methodOnInboxOpenCardsSync = 'onInboxOpenCardsSync';
 const String methodPullToRefreshCardsSync = 'onPullToRefreshCardsSync';
-const String methodOnAppOpenCardsSync = 'onAppOpenCardsSync';
+const String methodOnCardsSync = 'onCardsSync';
 
 //JSON Values Constants
 const String argumentPullToRefreshSync = 'PULL_TO_REFRESH';
 const String argumentInboxOpenSync = 'INBOX_OPEN';
 const String argumentAppOpenSync = 'APP_OPEN';
 const String argumentAllCards = 'All';
+const String argumentImmediateSync = 'IMMEDIATE';